Refactoring following transfer of methods to PyOptiSLang.
