+++ src/ansys/templates/python/osl_solution/{{cookiecutter.__project_name_slug}}/src/ansys/solutions/{{cookiecutter.__solution_name_slug}}/ui/app.py
@@ -2,12 +2,12 @@
 
 """Application."""
 
+import dash_bootstrap_components as dbc
+import dash_uploader as du
 import os
 import tempfile
 
-import dash_bootstrap_components as dbc
 from dash_extensions.enrich import DashProxy, MultiplexerTransform, NoOutputTransform, TriggerTransform
-import dash_uploader as du
 
 
 app = DashProxy(
