+++ src/ansys/templates/python/osl_solution/{{cookiecutter.__project_name_slug}}/src/ansys/solutions/{{cookiecutter.__solution_name_slug}}/ui/components/commands.py
@@ -1,13 +1,13 @@
 # ©2023, ANSYS Inc. Unauthorized use, distribution or duplication is prohibited.
 
 import dash_bootstrap_components as dbc
-from dash.exceptions import PreventUpdate
-from dash_extensions.enrich import Output, Input, State, html, dcc, callback, MATCH, callback_context
 import uuid
 
+from dash_extensions.enrich import Output, Input, State, html, dcc, callback, MATCH, callback_context
+
 from ansys.saf.glow.client.dashclient import DashClient
-from ansys.solutions.{{ cookiecutter.__solution_name_slug }}.solution.problem_setup_step import ProblemSetupStep
 from ansys.solutions.{{ cookiecutter.__solution_name_slug }}.solution.definition import {{ cookiecutter.__solution_definition_name }}
+from ansys.solutions.{{ cookiecutter.__solution_name_slug }}.solution.problem_setup_step import ProblemSetupStep
 
 
 class ProjectCommandsAIO(html.Div):
@@ -200,6 +200,8 @@ class ProjectCommandsAIO(html.Div):
         project = DashClient[{{ cookiecutter.__solution_definition_name }}].get_project(pathname)
         problem_setup_step = project.steps.problem_setup_step
 
+        problem_setup_step.selected_actor_from_command = problem_setup_step.selected_actor_from_treeview
+
         if "restart_button" in triggered_id:
             problem_setup_step.restart()
         elif "stop_gently_button" in triggered_id:
@@ -387,6 +389,8 @@ class ActorCommandsAIO(html.Div):
 
         project = DashClient[{{ cookiecutter.__solution_definition_name }}].get_project(pathname)
         problem_setup_step = project.steps.problem_setup_step
+
+        problem_setup_step.selected_actor_from_command = problem_setup_step.selected_actor_from_treeview
 
         if "restart_button" in triggered_id:
             problem_setup_step.restart()
