+++ src/ansys/templates/python/osl_solution/{{cookiecutter.__project_name_slug}}/src/ansys/solutions/{{cookiecutter.__solution_name_slug}}/ui/views/scenery_view.py
@@ -2,8 +2,10 @@
 
 """Frontend of the scenery view."""
 
-from dash_extensions.enrich import html
 import optislang_dash_lib
+
+from dash_extensions.enrich import html
+
 from ansys.solutions.{{ cookiecutter.__solution_name_slug }}.solution.problem_setup_step import ProblemSetupStep
