+++ src/ansys/templates/python/osl_solution/{{cookiecutter.__project_name_slug}}/src/ansys/solutions/{{cookiecutter.__solution_name_slug}}/ui/views/visualization_view.py
@@ -2,10 +2,11 @@
 
 """Frontend of the visualization view."""
 
-from dash_extensions.enrich import html
 import optislang_dash_lib
-from ansys.solutions.{{ cookiecutter.__solution_name_slug }}.solution.problem_setup_step import ProblemSetupStep
 
+from dash_extensions.enrich import html
+
+from ansys.solutions.{{ cookiecutter.__solution_name_slug }}.solution.problem_setup_step import ProblemSetupStep
 
 
 def layout(problem_setup_step: ProblemSetupStep) -> html.Div:
