+++ tests/tests_templates/test_python_templates.py
@@ -366,11 +366,7 @@ OSL_SOLUTION_STRUCTURE = [
     "doc/Makefile",
     "examples/README.md",
     f"src/ansys/solutions/{OSL_SOLUTION_VARS['__solution_name_slug']}/model/assets/README.md",
-    f"src/ansys/solutions/{OSL_SOLUTION_VARS['__solution_name_slug']}/model/optislang/project_tree.py",
-    f"src/ansys/solutions/{OSL_SOLUTION_VARS['__solution_name_slug']}/model/optislang/server_commands.py",
-    f"src/ansys/solutions/{OSL_SOLUTION_VARS['__solution_name_slug']}/model/utils.py",
     f"src/ansys/solutions/{OSL_SOLUTION_VARS['__solution_name_slug']}/solution/definition.py",
-    f"src/ansys/solutions/{OSL_SOLUTION_VARS['__solution_name_slug']}/solution/monitoring_step.py",
     f"src/ansys/solutions/{OSL_SOLUTION_VARS['__solution_name_slug']}/solution/problem_setup_step.py",
     f"src/ansys/solutions/{OSL_SOLUTION_VARS['__solution_name_slug']}/ui/assets/css/style.css",
     f"src/ansys/solutions/{OSL_SOLUTION_VARS['__solution_name_slug']}/ui/assets/images/README.md",
@@ -388,11 +384,6 @@ OSL_SOLUTION_STRUCTURE = [
     f"src/ansys/solutions/{OSL_SOLUTION_VARS['__solution_name_slug']}/ui/pages/monitoring_page.py",
     f"src/ansys/solutions/{OSL_SOLUTION_VARS['__solution_name_slug']}/ui/pages/page.py",
     f"src/ansys/solutions/{OSL_SOLUTION_VARS['__solution_name_slug']}/ui/pages/problem_setup_page.py",
-    f"src/ansys/solutions/{OSL_SOLUTION_VARS['__solution_name_slug']}/ui/utils/alerts.py",
-    f"src/ansys/solutions/{OSL_SOLUTION_VARS['__solution_name_slug']}/ui/utils/common_functions.py",
-    f"src/ansys/solutions/{OSL_SOLUTION_VARS['__solution_name_slug']}/ui/utils/constants.py",
-    f"src/ansys/solutions/{OSL_SOLUTION_VARS['__solution_name_slug']}/ui/utils/monitoring.py",
-    f"src/ansys/solutions/{OSL_SOLUTION_VARS['__solution_name_slug']}/ui/utils/placeholders.py",
     f"src/ansys/solutions/{OSL_SOLUTION_VARS['__solution_name_slug']}/ui/views/design_table_view.py",
     f"src/ansys/solutions/{OSL_SOLUTION_VARS['__solution_name_slug']}/ui/views/project_summary_view.py",
     f"src/ansys/solutions/{OSL_SOLUTION_VARS['__solution_name_slug']}/ui/views/result_files_view.py",
@@ -401,6 +392,7 @@ OSL_SOLUTION_STRUCTURE = [
     f"src/ansys/solutions/{OSL_SOLUTION_VARS['__solution_name_slug']}/ui/views/summary_view.py",
     f"src/ansys/solutions/{OSL_SOLUTION_VARS['__solution_name_slug']}/ui/views/visualization_view.py",
     f"src/ansys/solutions/{OSL_SOLUTION_VARS['__solution_name_slug']}/ui/app.py",
+    f"src/ansys/solutions/{OSL_SOLUTION_VARS['__solution_name_slug']}/utilities/common_functions.py",
     f"src/ansys/solutions/{OSL_SOLUTION_VARS['__solution_name_slug']}/__init__.py",
     f"src/ansys/solutions/{OSL_SOLUTION_VARS['__solution_name_slug']}/main.py",
     "telemetry/grafana/compose.yaml",
